+++ CHANGES.md
@@ -12,7 +12,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
   *  Added "Junit Xml File" setting to the Gut control to specify the file.  "Junit Xml Timestamp" will include an epoch timestamp in the filename.
   *  `-gjunit_xml_file` and `-gjunit_xml_timestamp` are supported on the command line.
   *   `junit_xml_file` and `junit_xml_timestamp` are supported in the `.gutconfig.json` file.
-* __Issue #266__ Introduced `double_parameter_override` and `clear_double_parameter_overrides` which allow you to specify the number of paramters a method has.  Useful when working with vararg methods or NativeScripts.  This addresses #246 and #252.
+* __Issue #266__ Introduced `double_parameter_override` and `clear_double_parameter_overrides` which allow you to specify the number of paramters a method has.  Useful when working with vararg methods or NativeScripts.  This addresses #246 and #252.  Overrides are automcatically cleared between scripts, not between tests.
 
 ## Bug Fixes
 * __Issue 268__ Add message when `assert_signal_emitted_with_parameters` is passed bad parameters.